refactor: move checkDuplicateTestCase
