+++ lib/rule-tester/rule-tester.js
@@ -18,8 +18,7 @@ const assert = require("node:assert"),
 	{ Config } = require("../config/config"),
 	{ Linter, SourceCodeFixer } = require("../linter"),
 	{ interpolate, getPlaceholderMatcher } = require("../linter/interpolate"),
-	stringify = require("json-stable-stringify-without-jsonify"),
-	{ validateTests } = require("./helpers");
+	{ validateTests, checkDuplicateTestCase } = require("./helpers");
 
 const { FlatConfigArray } = require("../config/flat-config-array");
 const {
@@ -31,7 +30,6 @@ const ajv = require("../shared/ajv")({ strictDefaults: true });
 
 const parserSymbol = Symbol.for("eslint.RuleTester.parser");
 const { ConfigArraySymbol } = require("@eslint/config-array");
-const { isSerializable } = require("../shared/serialization");
 
 const jslang = require("../languages/js");
 const { SourceCode } = require("../languages/js/source-code");
@@ -145,11 +143,6 @@ const suggestionObjectParameters = new Set([
 	"output",
 ]);
 const friendlySuggestionObjectParameterList = `[${[...suggestionObjectParameters].map(key => `'${key}'`).join(", ")}]`;
-
-/*
- * Ignored test case properties when checking for test case duplicates.
- */
-const duplicationIgnoredParameters = new Set(["name", "errors", "output"]);
 
 const forbiddenMethods = [
 	"applyInlineConfig",
@@ -903,41 +896,6 @@ class RuleTester {
 			if (!equal(beforeAST, afterAST)) {
 				assert.fail("Rule should not modify AST.");
 			}
-		}
-
-		/**
-		 * Check if this test case is a duplicate of one we have seen before.
-		 * @param {string|Object} item test case object
-		 * @param {Set<string>} seenTestCases set of serialized test cases we have seen so far (managed by this function)
-		 * @returns {void}
-		 * @private
-		 */
-		function checkDuplicateTestCase(item, seenTestCases) {
-			if (!isSerializable(item)) {
-				/*
-				 * If we can't serialize a test case (because it contains a function, RegExp, etc), skip the check.
-				 * This might happen with properties like: options, plugins, settings, languageOptions.parser, languageOptions.parserOptions.
-				 */
-				return;
-			}
-
-			const normalizedItem =
-				typeof item === "string" ? { code: item } : item;
-			const serializedTestCase = stringify(normalizedItem, {
-				replacer(key, value) {
-					// "this" is the currently stringified object --> only ignore top-level properties
-					return normalizedItem !== this ||
-						!duplicationIgnoredParameters.has(key)
-						? value
-						: void 0;
-				},
-			});
-
-			assert(
-				!seenTestCases.has(serializedTestCase),
-				"detected duplicate test case",
-			);
-			seenTestCases.add(serializedTestCase);
 		}
 
 		/**
